Apply global namespaces to files and rename for consistency.
