+++ ChatServer/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "ChatServer": {
       "commandName": "Project",
-      "commandLineArgs": "192.168.1.13 50000 5"
+      "commandLineArgs": "127.0.0.1 50000 5"
     }
   }
 }
