+++ SharedClasses/UserUpdates.cs
@@ -2,42 +2,41 @@
 using System.Collections.Generic;
 using System.Linq;
 
-namespace ChatModel
+namespace ChatModel;
+
+/// <summary>
+/// Class representing collection of updates to all user's conversations.
+/// </summary>
+public class UserUpdates : IEnumerable<ConversationUpdates>
 {
-	/// <summary>
-	/// Class representing collection of updates to all user's conversations.
-	/// </summary>
-	public class UserUpdates : IEnumerable<ConversationUpdates>
+	private List<ConversationUpdates> conversations;
+
+	public UserUpdates()
 	{
-		private List<ConversationUpdates> conversations;
+		this.conversations = new List<ConversationUpdates>();
+	}
+	public UserUpdates(ConversationUpdates[] conversations)
+	{
+		this.conversations = conversations.ToList<ConversationUpdates>();
+	}
 
-		public UserUpdates()
-		{
-			this.conversations = new List<ConversationUpdates>();
-		}
-		public UserUpdates(ConversationUpdates[] conversations)
-		{
-			this.conversations = conversations.ToList<ConversationUpdates>();
-		}
+	/// <summary>
+	/// Adds conversation updates object to the collection.
+	/// </summary>
+	/// <param name="conversation">Conversation updates to add.</param>
+	public void addConversation(ConversationUpdates conversation)
+	{
+		this.conversations.Add(conversation);
+	}
 
-		/// <summary>
-		/// Adds conversation updates object to the collection.
-		/// </summary>
-		/// <param name="conversation">Conversation updates to add.</param>
-		public void addConversation(ConversationUpdates conversation)
-		{
-			this.conversations.Add(conversation);
-		}
+	IEnumerator IEnumerable.GetEnumerator()
+	{
+		return ((IEnumerable)conversations).GetEnumerator();
+	}
 
-		IEnumerator IEnumerable.GetEnumerator()
-		{
-			return ((IEnumerable)conversations).GetEnumerator();
-		}
-
-		public IEnumerator<ConversationUpdates> GetEnumerator()
-		{
-			return ((IEnumerable<ConversationUpdates>)conversations).GetEnumerator();
-		}
+	public IEnumerator<ConversationUpdates> GetEnumerator()
+	{
+		return ((IEnumerable<ConversationUpdates>)conversations).GetEnumerator();
 	}
 }
