+++ graph-chat-app/ViewModel/ConversationCollectionViewModel.cs
@@ -4,36 +4,35 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Windows.Input;
 
-namespace GraphChatApp.ViewModel
+namespace GraphChatApp.ViewModel;
+
+class ConversationCollectionViewModel : ViewModel
 {
-	class ConversationCollectionViewModel : ViewModel
+	private ObservableCollection<ConversationViewModel> conversations;
+	private ChatSystem chatSystem;
+	private Action<Conversation> enterConversation;
+
+	public ConversationCollectionViewModel(ChatSystem chatSystem, Action<Conversation> enterConversation)
 	{
-		private ObservableCollection<ConversationViewModel> conversations;
-		private ChatSystem chatSystem;
-		private Action<Conversation> enterConversation;
+		this.chatSystem = chatSystem;
+		chatSystem.PropertyChanged += OnPropertyChanged;
+		this.enterConversation = enterConversation;
+		this.conversations = new ObservableCollection<ConversationViewModel>(
+			chatSystem.Conversations.Select(conv => new ConversationViewModel(conv.Value, enterConversation)).ToList()
+		);
+	}
 
-		public ConversationCollectionViewModel(ChatSystem chatSystem, Action<Conversation> enterConversation)
+	public ObservableCollection<ConversationViewModel> observableConversations
+	{
+		get
 		{
-			this.chatSystem = chatSystem;
-			chatSystem.PropertyChanged += OnPropertyChanged;
-			this.enterConversation = enterConversation;
-			this.conversations = new ObservableCollection<ConversationViewModel>(
-					chatSystem.Conversations.Select(conv => new ConversationViewModel(conv.Value, enterConversation)).ToList()
-				);
-		}
-
-		public ObservableCollection<ConversationViewModel> observableConversations
-		{
-			get
+			if (chatSystem.Conversations.Count != conversations.Count)
 			{
-				if (chatSystem.Conversations.Count != conversations.Count)
-				{
-					this.conversations = new ObservableCollection<ConversationViewModel>(
-							chatSystem.ObservableConversations.Select(conv => new ConversationViewModel(conv, enterConversation)).ToList()
-						);
-				}
-				return conversations;
+				this.conversations = new ObservableCollection<ConversationViewModel>(
+					chatSystem.ObservableConversations.Select(conv => new ConversationViewModel(conv, enterConversation)).ToList()
+				);
 			}
+			return conversations;
 		}
 	}
 }
