+++ graph-chat-app/ViewModel/ConversationCreatorViewModel.cs
@@ -5,124 +5,123 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
 
-namespace GraphChatApp.ViewModel
+namespace GraphChatApp.ViewModel;
+
+class CreateConversationCommand : ICommand
 {
-	class CreateConversationCommand : ICommand
+	public event EventHandler CanExecuteChanged = (sender, e) => { };
+	private Func<bool> canExecute;
+	private Action execute;
+	public CreateConversationCommand(Func<bool> canExecute, Action execute)
 	{
-		public event EventHandler CanExecuteChanged = (sender, e) => { };
-		private Func<bool> canExecute;
-		private Action execute;
-		public CreateConversationCommand(Func<bool> canExecute, Action execute)
-		{
-			this.canExecute = canExecute;
-			this.execute = execute;
-		}
+		this.canExecute = canExecute;
+		this.execute = execute;
+	}
 
-		public void InvokeCanExecuteChanged()
-		{
-			CanExecuteChanged(this, new());
-		}
+	public void InvokeCanExecuteChanged()
+	{
+		CanExecuteChanged(this, new());
+	}
 
-		public bool CanExecute(object parameter)
-		{
-			return canExecute();
-		}
+	public bool CanExecute(object parameter)
+	{
+		return canExecute();
+	}
 
-		public void Execute(object parameter)
+	public void Execute(object parameter)
+	{
+		execute();
+	}
+}
+class ConversationCreatorViewModel : ViewModel
+{
+	MainWindow window;
+
+	string csvusers;
+	string conversationName;
+
+	/// <summary>
+	/// A user-inputed string representing comma-separated usernames
+	/// </summary>
+	public string CSVUsers
+	{
+		get
 		{
-			execute();
+			return csvusers;
+		}
+		set
+		{
+			csvusers = value;
+			CreateConversation.InvokeCanExecuteChanged();
 		}
 	}
-	class ConversationCreatorViewModel : ViewModel
+
+	/// <summary>
+	/// Proposed name of conversation to be added
+	/// </summary>
+	public string ConversationName
 	{
-		MainWindow window;
-
-		string csvusers;
-		string conversationName;
-
-		/// <summary>
-		/// A user-inputed string representing comma-separated usernames
-		/// </summary>
-		public string CSVUsers
+		get
 		{
-			get
+			return conversationName;
+		}
+		set
+		{
+			conversationName = value;
+			CreateConversation.InvokeCanExecuteChanged();
+		}
+	}
+
+	/// <summary>
+	/// Names of users to be added to new conversation
+	/// </summary>
+	List<string> usernames
+	{
+		get
+		{
+			var currentUser = window.app.Client.ChatSystem.getUserName();
+			var users = CSVUsers
+				.Split(new char[] { ',', '\n' }, StringSplitOptions.RemoveEmptyEntries)
+				.Select(user => user.Trim())
+				.Where(user => !String.IsNullOrWhiteSpace(user))
+				.ToList();
+			if (!users.Contains(currentUser))
 			{
-				return csvusers;
+				users.Add(currentUser);
 			}
-			set
-			{
-				csvusers = value;
-				CreateConversation.InvokeCanExecuteChanged();
-			}
+			return users;
 		}
+	}
+	/// <summary>
+	/// Command for creating new conversation to be bound to a button
+	/// </summary>
+	public CreateConversationCommand CreateConversation { get; set; }
 
-		/// <summary>
-		/// Proposed name of conversation to be added
-		/// </summary>
-		public string ConversationName
-		{
-			get
-			{
-				return conversationName;
-			}
-			set
-			{
-				conversationName = value;
-				CreateConversation.InvokeCanExecuteChanged();
-			}
-		}
+	public ConversationCreatorViewModel(MainWindow window)
+	{
+		this.window = window;
+		CreateConversation = new CreateConversationCommand(canCreate, Create);
+	}
 
-		/// <summary>
-		/// Names of users to be added to new conversation
-		/// </summary>
-		List<string> usernames
-		{
-			get
-			{
-				var currentUser = window.app.Client.ChatSystem.getUserName();
-				var users = CSVUsers
-						.Split(new char[] { ',', '\n' }, StringSplitOptions.RemoveEmptyEntries)
-						.Select(user => user.Trim())
-						.Where(user => !String.IsNullOrWhiteSpace(user))
-						.ToList();
-				if (!users.Contains(currentUser))
-				{
-					users.Add(currentUser);
-				}
-				return users;
-			}
-		}
-		/// <summary>
-		/// Command for creating new conversation to be bound to a button
-		/// </summary>
-		public CreateConversationCommand CreateConversation { get; set; }
+	/// <summary>
+	/// Decides, wheter new Conversation can be created
+	/// </summary>
+	/// <returns></returns>
+	bool canCreate()
+	{
+		return !(String.IsNullOrWhiteSpace(CSVUsers) && String.IsNullOrWhiteSpace(conversationName));
+	}
 
-		public ConversationCreatorViewModel(MainWindow window)
-		{
-			this.window = window;
-			CreateConversation = new CreateConversationCommand(canCreate, Create);
-		}
-
-		/// <summary>
-		/// Decides, wheter new Conversation can be created
-		/// </summary>
-		/// <returns></returns>
-		bool canCreate()
-		{
-			return !(String.IsNullOrWhiteSpace(CSVUsers) && String.IsNullOrWhiteSpace(conversationName));
-		}
-
-		/// <summary>
-		/// Creates new conversation in system, and sends the info to the server
-		/// </summary>
-		void Create()
-		{
-			window.OnConversationAdded(
-				conversationName, 
-				usernames.ToArray());
-			CSVUsers = "";
-			ConversationName = "";
-			window.MainFrame.NavigationService.GoBack();
-		}
+	/// <summary>
+	/// Creates new conversation in system, and sends the info to the server
+	/// </summary>
+	void Create()
+	{
+		window.OnConversationAdded(
+			conversationName, 
+			usernames.ToArray());
+		CSVUsers = "";
+		ConversationName = "";
+		window.MainFrame.NavigationService.GoBack();
 	}
 }
