+++ graph-chat-app/ViewModel/ConversationViewModel.cs
@@ -2,44 +2,43 @@
 using System;
 using System.Windows.Input;
 
-namespace GraphChatApp.ViewModel
+namespace GraphChatApp.ViewModel;
+
+class EnterConversationCommand : ICommand
 {
-	class EnterConversationCommand : ICommand
+	public event EventHandler CanExecuteChanged = (sender, e) => { };
+	private Action execute;
+	public EnterConversationCommand(Action execute)
 	{
-		public event EventHandler CanExecuteChanged = (sender, e) => { };
-		private Action execute;
-		public EnterConversationCommand(Action execute)
-		{
-			this.execute = execute;
-		}
-		public bool CanExecute(object parameter) => true;
-		public void Execute(object parameter) => execute();
+		this.execute = execute;
 	}
-	class ConversationViewModel : ViewModel
+	public bool CanExecute(object parameter) => true;
+	public void Execute(object parameter) => execute();
+}
+class ConversationViewModel : ViewModel
+{
+	private Conversation conversation;
+	private Action<Conversation> enteringMethod;
+
+	public ConversationViewModel(Conversation conversation, Action<Conversation> action)
 	{
-		private Conversation conversation;
-		private Action<Conversation> enteringMethod;
+		this.conversation = conversation;
+		conversation.PropertyChanged += OnPropertyChanged;
+		this.enteringMethod = action;
+		EnterCommand = new EnterConversationCommand(EnterConversation);
+	}
 
-		public ConversationViewModel(Conversation conversation, Action<Conversation> action)
+	public string Name
+	{
+		get
 		{
-			this.conversation = conversation;
-			conversation.PropertyChanged += OnPropertyChanged;
-			this.enteringMethod = action;
-			EnterCommand = new EnterConversationCommand(EnterConversation);
+			return conversation.Name;
 		}
+	}
 
-		public string Name
-		{
-			get
-			{
-				return conversation.Name;
-			}
-		}
-
-		public EnterConversationCommand EnterCommand { get; set; }
-		void EnterConversation()
-		{
-			enteringMethod(conversation);
-		}
+	public EnterConversationCommand EnterCommand { get; set; }
+	void EnterConversation()
+	{
+		enteringMethod(conversation);
 	}
 }
