--- graph-chat-app/ViewModel/MessageViewModelBase.cs
@@ -1,28 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Windows;
-using System.Windows.Data;
-
-namespace GraphChatApp.ViewModel
-{
-	public abstract class MessageViewModelBase : ViewModel
-	{
-		protected ObservableCollection<MessageViewModelBase> childrenMessages;
-		protected const float messageWidth = 100;
-		protected const float messageHeight = 80;
-		protected const float messageMargin = 10;
-		public bool IsWrite { get; } = false;
-
-		public Visibility ChildrenVisibility => !IsWrite ? Visibility.Visible : Visibility.Collapsed;
-		public Visibility EditorVisibility => !IsWrite ? Visibility.Collapsed : Visibility.Visible;
-
-		public abstract string Author { get; }
-		public abstract string tempText { get; set; }
-
-		public virtual bool CanShowMessageEditor
-		{
-			get { return false; }
-		}
-	}
-}
